+++ openfasoc/generators/gdsfactory-gen/glayout/placement/two_transistor_interdigitized.py
@@ -1,19 +1,19 @@
 from glayout.pdk.mappedpdk import MappedPDK
 from pydantic import validate_arguments
 from gdsfactory.component import Component
-from glayout.fet import nmos, pmos, multiplier
+from glayout.primitives.fet import nmos, pmos, multiplier
 from glayout.pdk.util.comp_utils import evaluate_bbox
 from typing import Literal, Union
 from glayout.pdk.util.port_utils import rename_ports_by_orientation, rename_ports_by_list
 from glayout.pdk.util.comp_utils import prec_ref_center
 from glayout.routing.straight_route import straight_route
 from gdsfactory.functions import transformed
-from glayout.guardring import tapring
+from glayout.primitives.guardring import tapring
 from glayout.pdk.util.port_utils import add_ports_perimeter
 from gdsfactory.cell import clear_cache
 
 
-#from glayout.common.two_transistor_interdigitized import two_nfet_interdigitized; from glayout.pdk.sky130_mapped import sky130_mapped_pdk as pdk; biasParams=[6,2,4]; rmult=2
+#from glayout.placement.two_transistor_interdigitized import two_nfet_interdigitized; from glayout.pdk.sky130_mapped import sky130_mapped_pdk as pdk; biasParams=[6,2,4]; rmult=2
 
 @validate_arguments
 def two_transistor_interdigitized(
@@ -32,7 +32,7 @@ def two_transistor_interdigitized(
     deviceA_and_B = the device to place for both transistors (either nfet or pfet)
     dummy = place dummy at the edges of the interdigitized place (true by default). you can specify tuple to place only on one side
     kwargs = key word arguments for device. 
-    ****NOTE: These are the same as glayout.fet.multiplier arguments EXCLUDING dummy, sd_route_extension, and pdk options
+    ****NOTE: These are the same as glayout.primitives.fet.multiplier arguments EXCLUDING dummy, sd_route_extension, and pdk options
     """
     if isinstance(dummy, bool):
         dummy = (dummy, dummy)
@@ -107,7 +107,7 @@ def two_nfet_interdigitized(
     numcols = a single col is actually one col for both nfets (so AB). 2 cols = ABAB ... so on
     dummy = place dummy at the edges of the interdigitized place (true by default). you can specify tuple to place only on one side
     kwargs = key word arguments for multiplier. 
-    ****NOTE: These are the same as glayout.fet.multiplier arguments EXCLUDING dummy, sd_route_extension, and pdk options
+    ****NOTE: These are the same as glayout.primitives.fet.multiplier arguments EXCLUDING dummy, sd_route_extension, and pdk options
     tie_layers: tuple[str,str] specifying (horizontal glayer, vertical glayer) or well tie ring. default=("met2","met1")
     """
     base_multiplier = two_transistor_interdigitized(pdk, numcols, "nfet", dummy, **kwargs)
