+++ openfasoc/generators/gdsfactory-gen/glayout/primitives/via_gen.py
@@ -270,42 +270,3 @@ def via_array(
     return component_snap_to_grid(rename_ports_by_orientation(viaarray))
 
 
-if __name__ == "__main__":
-    from .pdk.util.standard_main import pdk, parser
-    from .pdk.util.custom_comp_utils import print_ports
-    from pathlib import Path
-
-    # default behavoir is to run one design and exit
-    parser.add_argument("--all", "-a", action="store_true", help="runs all tests")
-    parser.add_argument("--viastack", "-s", action="store_true", help="runs all via_stack tests")
-    parser.add_argument("--viaarray", "-v", action="store_true", help="runs all via_array tests")
-    parser.add_argument("--write", "-w", help="writes all gds files to directory specfied")
-    parser.add_argument("--ports", action="store_true", help="print ports")
-    args = parser.parse_args()
-    # run comps
-    comps = list()
-    if args.viaarray or args.all:
-        layers = ["poly", "met1", "met2", "met3"]
-        for lay1 in layers:
-            for lay2 in layers:
-                comps.append(via_array(pdk, lay1, lay2, lay_bottom=True))
-    elif args.viastack or args.all:
-        layers = ["poly", "met1", "met2", "met3"]
-        for lay1 in layers:
-            for lay2 in layers:
-                comps.append(via_stack(pdk, lay1, lay2,fullbottom=True,fulltop=True))
-    else:
-        myarray = via_array(pdk, "poly", "met2",size=(5,4))
-    # show and write (if write is specfied)
-    if args.write:
-        gds_write_path = Path(args.write)
-        if not gds_write_path.is_dir():
-            raise ValueError("gds write must be a dir path")
-        for comp in comps:
-            comp.write_gds(comp.name+".gds")
-    for comp in comps:
-        comp.show()
-    # print_ports
-    if args.ports:
-        for comp in comps:
-            print_ports(myarray)
