+++ openfasoc/generators/gdsfactory-gen/glayout/routing/L_route.py
@@ -3,7 +3,7 @@ from gdsfactory.component import Component
 from gdsfactory.port import Port
 from glayout.pdk.mappedpdk import MappedPDK
 from typing import Optional, Union
-from glayout.via_gen import via_stack, via_array
+from glayout.primitives.via_gen import via_stack, via_array
 from glayout.pdk.util.comp_utils import evaluate_bbox, align_comp_to_port, to_decimal, to_float, prec_ref_center, get_primitive_rectangle
 from glayout.pdk.util.port_utils import rename_ports_by_orientation, rename_ports_by_list, print_ports, assert_port_manhattan, assert_ports_perpindicular
 from decimal import Decimal
@@ -107,14 +107,3 @@ def L_route(
 	return rename_ports_by_orientation(Lroute.flatten())
 
 
-if __name__ == "__main__":
-	from glayout.pdk.util.standard_main import pdk
-	
-	routebetweentop = get_primitive_rectangle(layer=pdk.get_glayer("met1"),size=(1,1)).ref()
-	routebetweentop.movey(-4).movex(7)
-	routebetweenbottom = get_primitive_rectangle(layer=pdk.get_glayer("met1"), size=(1, 0.5))
-	mycomp = L_route(pdk,routebetweentop.ports["e4"],routebetweenbottom.ports["e1"])
-	mycomp.unlock()
-	mycomp.add(routebetweentop)
-	mycomp << routebetweenbottom
-	mycomp.flatten().show()
